Ducpmph33321new

+++ app/Http/Controllers/Backend/Account/AdminController.php
@@ -48,6 +48,7 @@ class AdminController extends Controller
      */
     public function index(AccountListRequest $request)
     {
+        $this->authorize('modules', '' . self::OBJECT . '.index');
         $this->tempImageService->deleteTempImagesForUser();
         // Validate the request data
         $request->validated();
@@ -168,7 +169,7 @@ class AdminController extends Controller
 
             return redirect()->back()->with('success', 'Admin delete successfully!');
         } catch (\Exception $e) {
-            
+
             return redirect()->back()->with('error', $e->getMessage());
         }
     }
